[Windows] Driver installer refactor
