+++ windows/QMK Toolbox/QMK Toolbox.csproj
@@ -100,6 +100,7 @@
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="Flashing.cs" />
+    <Compile Include="Helpers\DriverInstaller.cs" />
     <Compile Include="Helpers\EmbeddedResourceHelper.cs" />
     <Compile Include="Info.cs" />
     <Compile Include="Printing.cs" />
